requirements.txt
Korpora==0.2.0
pandas==2.0.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
translate-toolkit==3.10.0
